embeddings/embed_and_store.py

import faiss
from sentence_transformers import SentenceTransformer
import numpy as np

def embed_texts(texts):
    model = SentenceTransformer('all-MiniLM-L6-v2')
    embeddings = model.encode(texts)
    index = faiss.IndexFlatL2(embeddings.shape[1])
    index.add(np.array(embeddings))
    return index
