requirements.txt

flask
faiss-cpu
sentence-transformers
nltk
schedule
